Simplify the routeguide project structure
