+++ routeguide/client/src/main/scala/implicits.scala
@@ -14,27 +14,20 @@
  * limitations under the License.
  */
 
-package example.routeguide.client.io
+package example.routeguide.client
 
-import cats.effect.IO
-import org.log4s._
-import example.routeguide.client.io.implicits._
-import example.routeguide.client.ClientProgram._
+import cats.effect.{IO, Resource}
+import example.routeguide.client.handlers.RouteGuideClientHandler
+import example.routeguide.protocol.service._
+import example.routeguide.runtime._
+import example.routeguide.client.runtime._
 
-object ClientAppIO {
+trait ClientIOImplicits extends RouteGuide with ClientConf {
 
-  def main(args: Array[String]): Unit = {
+  implicit val routeGuideServiceClient: Resource[IO, RouteGuideService[IO]] =
+    RouteGuideService.client[IO](channelFor)
 
-    val logger = getLogger
-
-    logger.info(s"${Thread.currentThread().getName} Starting client, interpreting to Future ...")
-
-    clientProgram[IO].unsafeRunSync()
-
-    logger.info(s"${Thread.currentThread().getName} Finishing program interpretation ...")
-
-    System.in.read()
-    (): Unit
-  }
-
+  implicit val routeGuideClientHandler: RouteGuideClientHandler[IO] =
+    new RouteGuideClientHandler[IO]
 }
+object implicits extends ClientIOImplicits
